requirements.txt
click==8.3.0
firebase_admin==7.1.0
Flask==3.1.2
ollama==0.6.0
openai==2.5.0
python-dotenv==1.1.1
waitress==3.0.2
